aider: Add vscode launch.json file with two configurations for recovering wallet-20121216-n.dat and wallet-20140324-n.dat.
# Aider chat conversation:

USER:
Add vscode launch.json for following command lines:

./pywallet.py --recover --recov_device=wallet-20121216-n.dat --recov_size=368640 --recov_outputdir=wallet-20121216-n.dat-rec
./pywallet.py --recover --recov_device=wallet-20140324-n.dat --recov_size=2277376 --recov_outputdir=wallet-20140324-n.dat-rec

ASSISTANT: To create a `launch.json` file for Visual Studio Code that includes the specified command lines, we will need to add a new file to the `.vscode` directory within the project. The `launch.json` file defines configurations for debugging.

Here are the *SEARCH/REPLACE blocks* for creating a new `launch.json` file:

.vscode/launch.json
```json
<<<<<<< SEARCH
=======
{
    "version": "0.2.0",
    "configurations": [
        {
            "name": "Recover wallet-20121216-n.dat",
            "type": "python",
            "request": "launch",
            "program": "${workspaceFolder}/pywallet.py",
            "args": [
                "--recover",
                "--recov_device=wallet-20121216-n.dat",
                "--recov_size=368640",
                "--recov_outputdir=wallet-20121216-n.dat-rec"
            ]
        },
        {
            "name": "Recover wallet-20140324-n.dat",
            "type": "python",
            "request": "launch",
            "program": "${workspaceFolder}/pywallet.py",
            "args": [
                "--recover",
                "--recov_device=wallet-20140324-n.dat",
                "--recov_size=2277376",
                "--recov_outputdir=wallet-20140324-n.dat-rec"
            ]
        }
    ]
}
>>>>>>> REPLACE
```

This will create a new `launch.json` file in the `.vscode` directory with two configurations, one for each of the recovery commands provided. When you open this project in Visual Studio Code, you will be able to run these configurations directly from the Run and Debug sidebar.